+++ test/functional/datasets/GetDatasetAvailableDatasetTypes.test.ts
@@ -14,16 +14,18 @@ describe('getDatasetAvailableDatasetTypes', () => {
 
     test('should return available dataset types', async () => {
       const actualDatasetTypes: DatasetType[] = await getDatasetAvailableDatasetTypes.execute()
-      const expectedDatasetTypes = {
-        id: 1,
-        name: 'dataset',
-        linkedMetadataBlocks: [],
-        availableLicenses: []
-      }
-
-      // check that the actual dataset types include the expected dataset types
-      // (without requiring an exact match, since other dataset types may be created by concurrent tests)
-      expect(actualDatasetTypes).toContainEqual(expectedDatasetTypes)
+      const expectedDatasetTypes = [
+        {
+          id: 1,
+          name: 'dataset',
+          linkedMetadataBlocks: [],
+          availableLicenses: [],
+          description:
+            'A study, experiment, set of observations, or publication. A dataset can comprise a single file or multiple files.',
+          displayName: 'Dataset'
+        }
+      ]
+      expect(actualDatasetTypes).toEqual(expectedDatasetTypes)
     })
   })
 })