+++ test/unit/datasets/GetDatasetAvailableDatasetTypes.test.ts
@@ -12,14 +12,19 @@ describe('GetDatasetAvailableDatasetTypes', () => {
         {
           id: 1,
           name: 'dataset',
+          displayName: 'Dataset',
           linkedMetadataBlocks: [],
-          availableLicenses: []
+          availableLicenses: [],
+          description:
+            'A study, experiment, set of observations, or publication. A dataset can comprise a single file or multiple files.'
         },
         {
           id: 2,
           name: 'software',
+          displayName: 'Software',
           linkedMetadataBlocks: ['codeMeta20'],
-          availableLicenses: ['MIT', 'Apache-2.0']
+          availableLicenses: ['MIT', 'Apache-2.0'],
+          description: 'Software data and metadata.'
         }
       ]